db_sz int에서 long으로 변경

+++ eXperDB-Management-WebConsole/src/main/java/com/k4m/dx/tcontrol/backup/service/WorkLogVO.java
@@ -37,7 +37,7 @@ public class WorkLogVO {
 	private String bck_wrk_id;
 	private String log_filenm;
 	private String backrest_gbn;
-	private int db_sz;
+	private long db_sz;
 
 	public String getBck_wrk_id() {
 		return bck_wrk_id;
@@ -327,11 +327,11 @@ public class WorkLogVO {
 		this.backrest_gbn = backrest_gbn;
 	}
 
-	public int getDb_sz() {
+	public long getDb_sz() {
 		return db_sz;
 	}
 
-	public void setDb_sz(int db_sz) {
+	public void setDb_sz(long db_sz) {
 		this.db_sz = db_sz;
 	}
